+++ CODEOWNERS
@@ -1,1 +1,1 @@
-* @seveibar @tscircuit/core
+* @seveibar
