+++ package.json
@@ -14,6 +14,7 @@
     "redaxios": "^0.5.1"
   },
   "devDependencies": {
+    "@tscircuit/fake-snippets": "^0.0.2",
     "@types/bun": "latest",
     "@types/node": "^22.10.5",
     "typescript": "^5.7.3"
