--- src/@types/global.d.ts
@@ -1,5 +1,0 @@
-declare global {
-    var axios: typeof import("axios").default;
-  }
-  
-  export {};
