+++ tests/fixtures/get-test-server.ts
@@ -1,0 +1,3 @@
+import { getTestServer } from "@tscircuit/fake-reddit/tests/fixtures/get-test-server";
+
+export { getTestServer };
